refactor overlay mixin to utils
