+++ application/lib/presentation/error/mixin/error_handling_manager_mixin.dart
@@ -1,8 +1,8 @@
 import 'package:xayn_discovery_app/domain/model/error/error_object.dart';
 import 'package:xayn_discovery_app/infrastructure/di/di_config.dart';
-import 'package:xayn_discovery_app/presentation/discovery_card/widget/overlay_data.dart';
-import 'package:xayn_discovery_app/presentation/discovery_card/widget/overlay_manager_mixin.dart';
 import 'package:xayn_discovery_app/presentation/error/widget/error_screen.dart';
+import 'package:xayn_discovery_app/presentation/utils/overlay/overlay_data.dart';
+import 'package:xayn_discovery_app/presentation/utils/overlay/overlay_manager_mixin.dart';
 import 'package:xayn_discovery_app/presentation/utils/tooltip_utils.dart';
 
 mixin ErrorHandlingManagerMixin<T> on OverlayManagerMixin<T> {
